Create types for ImmunizationFormData, FHIRImmunizationResource, FHIRBundle
